Implemented user debugging support for x86_64.
Reused x86 arch_user_debugger.cpp, with a few minor changes to make
the code work for both 32 and 64 bit. Something isn't quite working
right, if a breakpoint is hit the kernel will hang. Other than that
everything appears to work correctly.

+++ headers/private/kernel/arch/x86/arch_cpu.h
@@ -263,8 +263,8 @@ typedef struct arch_cpu_info {
 
 	struct X86PagingStructures* active_paging_structures;
 
-	uint32				dr6;	// temporary storage for debug registers (cf.
-	uint32				dr7;	// x86_exit_user_debug_at_kernel_entry())
+	size_t				dr6;	// temporary storage for debug registers (cf.
+	size_t				dr7;	// x86_exit_user_debug_at_kernel_entry())
 
 	// local TSS for this cpu
 	struct tss			tss;
